backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
pypdf==3.17.1
python-docx==1.1.0
docx2txt==0.8
pandas==2.1.3
openpyxl==3.1.2
docxtpl==0.20.2
aiofiles==23.2.1
openai==1.3.6
qdrant-client==1.6.9
numpy==1.26.4
scikit-learn==1.4.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0